+++ HOTFOLDER/DRAFT/OTMM_BATCH_1572627598804.xml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE TEAMS_ASSET_FILE [
+<!ENTITY file_1 SYSTEM " 20191024113759\PRS_H7_12972_XVG_12V_9287601267PL_DRAFT.pdf " NDATA application_pdf>
+<!ENTITY file_2 SYSTEM " 20191024113759\PRS_999999999902_DRAFT.pdf " NDATA application_pdf>
+<!ENTITY file_3 SYSTEM " 20191024113759\PRS_HB3fit_12359_RA_12V_9230025171PL_DRAFT.pdf " NDATA application_pdf>
+]>
+<TEAMS_ASSET_FILE>
+  <ASSETS>
+    <ASSET>
+      <METADATA>
+        <UOIS AUTHOR="PRINTSUITE" MODEL_ID="105" SUBJECT="9287601267PL">
+          <LUM_MD_FILE_INFO ASSET_OWNER="PMM^Aachen"/>
+          <LUM_MD_DIGI_HUB_INFO WCMS_CONFIDENTIALITY="Not Publishing"/>
+          <LUM_MD_MEDIA_INFO_TAB REGIONS="Independent"/>
+          <LUM_MD_MEDIA_INFO ASSET_TYPE="PRS" BRAND="Lumileds" KEYWORDS="DRAFT"/>
+          <SECURITY_POLICY_UOIS SEC_POLICY_ID="18"/>
+        </UOIS>
+      </METADATA>
+      <CONTENT>
+        <MASTER FILE="file_1"/>
+      </CONTENT>
+    </ASSET>
+    <ASSET>
+      <METADATA>
+        <UOIS AUTHOR="PRINTSUITE" MODEL_ID="105" SUBJECT="999999999902">
+          <LUM_MD_FILE_INFO ASSET_OWNER="PMM^Aachen"/>
+          <LUM_MD_DIGI_HUB_INFO WCMS_CONFIDENTIALITY="Not Publishing"/>
+          <LUM_MD_MEDIA_INFO_TAB REGIONS="Independent"/>
+          <LUM_MD_MEDIA_INFO ASSET_TYPE="PRS" BRAND="Lumileds" KEYWORDS="DRAFT"/>
+          <SECURITY_POLICY_UOIS SEC_POLICY_ID="18"/>
+        </UOIS>
+      </METADATA>
+      <CONTENT>
+        <MASTER FILE="file_2"/>
+      </CONTENT>
+    </ASSET>
+    <ASSET>
+      <METADATA>
+        <UOIS AUTHOR="PRINTSUITE" MODEL_ID="105" SUBJECT="9230025171PL">
+          <LUM_MD_FILE_INFO ASSET_OWNER="PMM^Aachen"/>
+          <LUM_MD_DIGI_HUB_INFO WCMS_CONFIDENTIALITY="Not Publishing"/>
+          <LUM_MD_MEDIA_INFO_TAB REGIONS="Independent"/>
+          <LUM_MD_MEDIA_INFO ASSET_TYPE="PRS" BRAND="Lumileds" KEYWORDS="DRAFT"/>
+          <SECURITY_POLICY_UOIS SEC_POLICY_ID="18"/>
+        </UOIS>
+      </METADATA>
+      <CONTENT>
+        <MASTER FILE="file_3"/>
+      </CONTENT>
+    </ASSET>
+  </ASSETS>
+</TEAMS_ASSET_FILE>
