+++ HOTFOLDER/bin/DraftExecutor.bat
@@ -1,3 +1,3 @@
 @ECHO OFF
-REM D:\OpenText\OTMM\MediaManagement\bin\bulkimport.bat -user tsuper -password OT@dmin321 -mode xml -dir D:\OpenText\OTMM\priint\Industrial -folder 98329293b9160c17cbefea7529c936b991bc1371 -nodelete -count 25 >> D:\OpenText\OTMM\HotFolder\logs\ps-industrial-asset-import.log
+D:\OpenText\OTMM\MediaManagement\bin\bulkimport.bat -user tsuper -password OT@dmin321 -mode xml -dir D:\eclipse-workspace\PriintPreProcessor\HOTFOLDER\DRAFT -folder 98329293b9160c17cbefea7529c936b991bc1371 -nodelete -count 25 >> D:\OpenText\OTMM\HotFolder\logs\ps-industrial-asset-import.log
 echo Executed the command
