--- priint/OTMM_BATCH_20191024113759.xml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<!DOCTYPE TEAMS_ASSET_FILE [
-<!ENTITY file1 SYSTEM "20191024113759\PRS_H7_12972_XVG_12V_9287601267PL_DRAFT.pdf" NDATA application_pdf>
-<!ENTITY file2 SYSTEM "20191024113759\PRS_999999999902_DRAFT.pdf" NDATA application_pdf>
-<!ENTITY file3 SYSTEM "20191024113759\PRS_HB3fit_12359_RA_12V_9230025171PL_DRAFT.pdf" NDATA application_pdf>
-]
->
-<TEAMS_ASSET_FILE>
-    <ASSETS>
-        <ASSET>
-            <METADATA>
-                <UOIS AUTHOR="PRINTSUITE" SUBJECT="9287601267PL" MODEL_ID="105" NAME="PRS_H7_12972_XVG_12V_9287601267PL_DRAFT.pdf">
-                    <LUM_MD_FILE_INFO ASSET_OWNER="PMM^Aachen"/>
-                    <LUM_MD_DIGI_HUB_INFO WCMS_CONFIDENTIALITY="Not Publishing"/>
-                    <LUM_MD_MEDIA_INFO_TAB REGIONS="Independent"/>
-                    <LUM_MD_MEDIA_INFO ASSET_TYPE="PRS" KEYWORDS="DRAFT" BRAND="Lumileds"/>
-                    <SECURITY_POLICY_UOIS SEC_POLICY_ID="18"/>
-                </UOIS>
-            </METADATA>
-            <CONTENT>
-                <MASTER FILE="file1" />
-            </CONTENT>
-        </ASSET>
-        <ASSET>
-            <METADATA>
-                <UOIS AUTHOR="PRINTSUITE" SUBJECT="999999999902" MODEL_ID="105" NAME="PRS_999999999902_DRAFT.pdf">
-                    <LUM_MD_FILE_INFO ASSET_OWNER="PMM^Aachen"/>
-                    <LUM_MD_DIGI_HUB_INFO WCMS_CONFIDENTIALITY="Not Publishing"/>
-                    <LUM_MD_MEDIA_INFO_TAB REGIONS="Independent"/>
-                    <LUM_MD_MEDIA_INFO ASSET_TYPE="PRS" KEYWORDS="DRAFT" BRAND="Lumileds"/>
-                    <SECURITY_POLICY_UOIS SEC_POLICY_ID="18"/>
-                </UOIS>
-            </METADATA>
-            <CONTENT>
-                <MASTER FILE="file2" />
-            </CONTENT>
-        </ASSET>
-        <ASSET>
-            <METADATA>
-                <UOIS AUTHOR="PRINTSUITE" SUBJECT="9230025171PL" MODEL_ID="105" NAME="PRS_HB3fit_12359_RA_12V_9230025171PL_DRAFT.pdf">
-                    <LUM_MD_FILE_INFO ASSET_OWNER="PMM^Aachen"/>
-                    <LUM_MD_DIGI_HUB_INFO WCMS_CONFIDENTIALITY="Not Publishing"/>
-                    <LUM_MD_MEDIA_INFO_TAB REGIONS="Independent"/>
-                    <LUM_MD_MEDIA_INFO ASSET_TYPE="PRS" KEYWORDS="DRAFT" BRAND="Lumileds"/>
-                    <SECURITY_POLICY_UOIS SEC_POLICY_ID="18"/>
-                </UOIS>
-            </METADATA>
-            <CONTENT>
-                <MASTER FILE="file3" />
-            </CONTENT>
-        </ASSET>
-    </ASSETS>
-</TEAMS_ASSET_FILE>
